ocr_pipeline/requirements.txt
pdf2image
httpx
boto3
Pillow
tqdm
python-dotenv
